we've separated the mime procces into its own function

+++ src/models/file.rs
@@ -6,12 +6,16 @@ use super::user::Role;
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct FileLog {
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub _id: Option<ObjectId>,
+
+    #[serde(with = "time::serde::rfc3339")]
     pub create_at: time::OffsetDateTime,
+
     pub file_name: String,
     pub channel: String,
     pub program_tv: String,
-    pub elapsed_upload: Option<usize>,
+    pub elapsed_upload: Option<u64>,
     pub owner: Owner,
     pub size: usize,
 }
